pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discord-pm"
version = "0.1.0"
description = "Discord bot for standup recording, transcription, and project management"
requires-python = ">=3.11"
dependencies = [
    "discord.py[voice]>=2.4",
    "openai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
